Create pipeline.sh
Adding the shell script to run the Python pipeline for creating our database. Please be patient while the script runs, as the JSON data downloaded is too large (148 MB for IT_20 and 75 MB for IPL), and will increase with further downloads as the website tracking this data keeps getting regular updates as more matches are played. The Python script takes nearly 15 to 20 minutes to parse through all the data and create the CSV databases.